restructure folders and components
