+++ src/App.js
@@ -1,7 +1,7 @@
 import React, {Component} from 'react';
 import logo from './logo.svg';
 import './App.css';
-import FrontPage from './sections/FrontPage/index.js';
+import FrontPage from './sections/FrontPage';
 
 class App extends Component {
   render() {
