+++ src/sections/FrontPage.js
@@ -1,5 +1,5 @@
 import React, {Component} from 'react';
-import './index.css';
+import './FrontPage.css';
 
 class FrontPage extends Component {
   render() {
