--- src/sections/Home/Home.js
@@ -1,14 +1,0 @@
-import React, { Component } from 'react';
-import './Home.css';
-
-class Home extends Component {
-  render() {
-    return (
-      <div className="Contact">
-        <h1>Contact Page</h1>
-      </div>
-    );
-  }
-}
-
-export default Home;
